Add files via upload
This a number guessing game where a random number is generated in a given range by the system each time the game is played. The number has to be guessed within limited number of guesses. If you are able to guess the number correctly within the guessing range you win otherwise system gives a message- Game Over.